# src/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.SaveConfigRequest:
    properties:
      baseChecksum:
        description: GET 回傳的 checksum，必填（檔案存在時）
        type: string
      config:
        type: string
    type: object
  handler.SaveConfigResponse:
    properties:
      backupPath:
        type: string
      message:
        type: string
    type: object
  handler.ServiceConfigResponse:
    properties:
      checksum:
        description: SHA-256 hex，供 PUT baseChecksum
        type: string
      config:
        type: string
      fragmentPath:
        type: string
      name:
        type: string
      size:
        type: integer
    type: object
...
  handler.conflictResponse:
    properties:
      currentChecksum:
        type: string
      error:
        type: string
    type: object
  handler.messageJSON:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  systemd.ValidateError:
    properties:
      line:
        type: integer
      message:
        type: string
    type: object
  systemd.ValidateResult:
    properties:
      available:
        type: boolean
      errors:
        items:
          $ref: '#/definitions/systemd.ValidateError'
        type: array
      message:
        type: string
      valid:
        type: boolean
    type: object
...
info:
  contact: {}
  title: Linux Service Manager API
  version: 1.0.0
paths:
...
  /login:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 以帳號密碼登入（form-urlencoded）。成功後設定 `session` cookie；亦有 rate limit（每
        IP 每分鐘 5 次嘗試）。
      parameters:
      - description: 使用者名稱
        in: formData
        name: username
        required: true
        type: string
      - description: 密碼
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{username, message}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 缺少帳號或密碼
          schema:
            $ref: '#/definitions/handler.messageJSON'
        "401":
          description: 帳號或密碼錯誤
          schema:
            $ref: '#/definitions/handler.messageJSON'
      summary: 登入
      tags:
      - Auth
...
  /services/{name}/config:
    get:
      description: 讀取指定服務的 systemd unit 設定檔內容（含 SHA-256 checksum 供後續 PUT baseChecksum）。鎖定服務亦可唯讀檢視。`read`
        scope Token 可用。
      parameters:
      - description: 服務名稱（systemd unit name）
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ServiceConfigResponse'
        "400":
          description: 服務名稱無效
          schema:
            $ref: '#/definitions/handler.messageJSON'
        "401":
          description: 未驗證
          schema:
            $ref: '#/definitions/handler.messageJSON'
        "404":
          description: 設定檔路徑或檔案不存在
          schema:
            $ref: '#/definitions/handler.messageJSON'
        "413":
          description: 設定檔過大
          schema:
            $ref: '#/definitions/handler.messageJSON'
        "500":
          description: 讀取失敗
          schema:
            $ref: '#/definitions/handler.messageJSON'
      security:
      - BearerAuth: []
      summary: 取得服務設定檔
      tags:
      - Service Config
    put:
      consumes:
      - application/json
      description: 覆寫指定服務的設定檔並執行 daemon-reload。需帶 GET 取得的 `baseChecksum` 做並發衝突偵測；衝突時回
        409 與 `currentChecksum`。需 `full` scope Token。\n\n**半成功**：設定檔已寫入但 daemon-reload
        失敗時回 500，附 `backupPath` 供手動還原。
      parameters:
      - description: 服務名稱（systemd unit name）
        in: path
        name: name
        required: true
        type: string
      - description: 新設定檔內容與 baseChecksum
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.SaveConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.SaveConfigResponse'
        "400":
          description: 請求無效或缺 base_checksum
          schema:
            $ref: '#/definitions/handler.messageJSON'
        "401":
          description: 未驗證
          schema:
            $ref: '#/definitions/handler.messageJSON'
        "403":
          description: 唯讀 Token 或路徑不可寫
          schema:
            $ref: '#/definitions/handler.messageJSON'
        "404":
          description: 設定檔路徑不存在
          schema:
            $ref: '#/definitions/handler.messageJSON'
        "409":
          description: 並發衝突（含 currentChecksum）
          schema:
            $ref: '#/definitions/handler.conflictResponse'
        "413":
          description: 設定檔超過 500KB
          schema:
            $ref: '#/definitions/handler.messageJSON'
        "500":
          description: 寫入/daemon-reload 失敗（含 backupPath）
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 儲存服務設定檔
      tags:
      - Service Config
  /services/{name}/config/validate:
    post:
      consumes:
      - application/json
      description: 以 `systemd-analyze verify` 驗證設定內容（不寫入）。systemd-analyze 不存在時回 200
        `{valid:false, available:false}`。`read` scope Token 可用。
      parameters:
      - description: 服務名稱（systemd unit name）
        in: path
        name: name
        required: true
        type: string
      - description: '{\'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/systemd.ValidateResult'
        "400":
          description: 服務名稱無效或請求無效
          schema:
            $ref: '#/definitions/handler.messageJSON'
        "401":
          description: 未驗證
          schema:
            $ref: '#/definitions/handler.messageJSON'
        "500":
          description: 無法建立暫存檔
          schema:
            $ref: '#/definitions/handler.messageJSON'
      security:
      - BearerAuth: []
      summary: 驗證服務設定檔
      tags:
      - Service Config
...
  /ws:
    get:
      description: '升級為 WebSocket 並即時推送服務狀態變更。**連線後不須傳入任何訊息** — 伺服器於狀態變更時主動推送快照（JSON：`{name,
        active, sub, unitFileState}`），連線建立時亦推送完整快照。\n\n**認證**：支援自訂 header 的 ws 客戶端請帶
        `Authorization: Bearer` header；瀏覽器原生 WebSocket 需 session cookie。\n**限制**：每個
        user 最多 5 個連線，超過回 close 1008（policy violation）。'
      responses:
        "101":
          description: Switching Protocols（之後為服務狀態快照 JSON TextMessage）
        "401":
          description: 未驗證
          schema:
            $ref: '#/definitions/handler.messageJSON'
      security:
      - BearerAuth: []
      summary: 服務狀態即時推送（WebSocket）
      tags:
      - WebSocket
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"